Fix a bunch of warnings
git-svn-id: svn://anonsvn.mit.edu/krb5/branches/mskrb-integ@21549 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/gssapi/krb5/k5unsealiov.c
@@ -378,7 +378,7 @@ kg_unseal_iov_token(OM_uint32 *minor_status,
 
     if (ctx->proto == 0)
 	code = kg_unseal_v1_iov(context, minor_status, ctx, iov, iov_count,
-				(ptr - (unsigned char *)header->buffer.value),
+				(size_t)(ptr - (unsigned char *)header->buffer.value),
 			        conf_state, qop_state, toktype);
     else
 	code = gss_krb5int_unseal_v3_iov(context, minor_status, ctx, iov, iov_count,
@@ -436,7 +436,7 @@ kg_unseal_stream_iov(OM_uint32 *minor_status,
 	goto cleanup;
     }
 
-    tiov = (gss_iov_buffer_desc *)calloc(iov_count + 2, sizeof(gss_iov_buffer_desc));
+    tiov = (gss_iov_buffer_desc *)calloc((size_t)iov_count + 2, sizeof(gss_iov_buffer_desc));
     if (tiov == NULL) {
 	code = ENOMEM;
 	goto cleanup;
